Functions, yay!
We now support function calls! We don't have support for forward
declaring functions right now though, so no mutual recursion is
possible.

The arguments are passed via the stack instead of through registers
(unlike the x86_64 calling convention, I think). We'll probably need
some sort of primitives built into the language for syscalls
eventually because of this.

Return types are also not checked, and right now it's possible to have
a function that doesn't return anything even when the caller expects it
to, error checking and reporting definitely needs to be improved.

+++ cup/common.h
@@ -3,4 +3,4 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-typedef int64_t i64;
+typedef long long int i64;
